Move PreallocateFile to fs package
