+++ internal/fs/preallocate_other.go
@@ -1,11 +1,11 @@
 //go:build !linux && !darwin
 // +build !linux,!darwin
 
-package restorer
+package fs
 
 import "os"
 
-func preallocateFile(wr *os.File, size int64) error {
+func PreallocateFile(wr *os.File, size int64) error {
 	// Maybe truncate can help?
 	// Windows: This calls SetEndOfFile which preallocates space on disk
 	return wr.Truncate(size)
