+++ internal/fs/preallocate_test.go
@@ -1,4 +1,4 @@
-package restorer
+package fs
 
 import (
 	"os"
@@ -7,7 +7,6 @@ import (
 	"syscall"
 	"testing"
 
-	"github.com/restic/restic/internal/fs"
 	"github.com/restic/restic/internal/test"
 )
 
@@ -23,7 +22,7 @@ func TestPreallocate(t *testing.T) {
 				test.OK(t, wr.Close())
 			}()
 
-			err = preallocateFile(wr, i)
+			err = PreallocateFile(wr, i)
 			if err == syscall.ENOTSUP {
 				t.SkipNow()
 			}
@@ -32,7 +31,7 @@ func TestPreallocate(t *testing.T) {
 			fi, err := wr.Stat()
 			test.OK(t, err)
 
-			efi := fs.ExtendedStat(fi)
+			efi := ExtendedStat(fi)
 			test.Assert(t, efi.Size == i || efi.Blocks > 0, "Preallocated size of %v, got size %v block %v", i, efi.Size, efi.Blocks)
 		})
 	}
